split standalone from index
